Move breakpoints utility to root of utilities folder
