+++ source/default/_patterns/00-protons/utilities/breakpoints.js
@@ -6,7 +6,7 @@ import enquire from 'enquire.js';
 
 import 'protons'; // Guarantees :root CSS variables referenced below
 
-import { sass2Array, cssVars2Obj } from '..';
+import { sass2Array, cssVars2Obj } from '.';
 
 import './_breakpoints.scss'; // Custom, e.g. --breakpoints: xs, sm, md, lg, xl;
